% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ibm_sent}
\alias{ibm_sent}
\title{Analyze some text for sentiment using IBM Watson.}
\usage{
ibm_sent(s_username, s_password, s_text)
}
\arguments{
\item{s_username}{Username.}

\item{s_password}{Password.}

\item{s_text}{Text to be analysed.}
}
\description{
This returns the result of the sentiment analysis on the provided text.
}
\examples{
ibm_sent("s_username", "s_password", "s_text")
ibm_sent("xxx-xxx-xxx", "abcdeabcde", "this is some text to analyse for sentiment")
}
\keyword{analysis}
\keyword{api}
\keyword{ibm}
\keyword{sentiment}
\keyword{tone}
\keyword{watson}
