% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ibm_tone}
\alias{ibm_tone}
\title{Analyze some text for tone using IBM Watson.}
\usage{
ibm_tone(t_username, t_password, t_text)
}
\arguments{
\item{t_username}{Username.}

\item{t_password}{Password.}

\item{t_text}{Text to be analysed.}
}
\description{
This returns the result of the tone analysis on the provided text.
}
\examples{
ibm_sent("t_username", "t_password", "t_text")
ibm_sent("xxx-xxx-xxx", "abcdeabcde", "this is some text to analyse for tone")
}
\keyword{analysis}
\keyword{api}
\keyword{ibm}
\keyword{sentiment}
\keyword{tone}
\keyword{watson}
